dev4/requirements.txt
Flask==2.0.1
importlib-metadata==4.7.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
pytesseract==0.3.8
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
